separate style from markup
